settings: storage area ID configuration
Enable the configuration of the ID of the flash area used to store
settings with FCB backend. The ID is not defined at compile time since
it can be useful to be able to specify a different area ID depending
on the runtime context. For example, if the settings are stored in
dual-bank flash memory that enables to swap the banks, the address of
the storage partition will typically not be the same depending on
whether or not the bank are swapped.

Signed-off-by: Thomas Altenbach <taltenbach@witekio.com>